fix: resolve Codacy file size violations by splitting large files
- Split CharacterValidationForm.test.tsx (492 lines) into 6 focused test files
- Split round-tracking-test-helpers.ts (472 lines) into 6 helper modules
- Split useRoundTracking.ts (460 lines) into 5 focused hook modules
- Split useRoundTracking.test.ts (455 lines) into 8 focused test files

All files now comply with 450-line limit while maintaining
comprehensive test coverage and functionality.
