+++ src/components/forms/character/__tests__/CharacterValidationForm.test.tsx
@@ -1,493 +1,22 @@
-import React from 'react';
-import { render, screen, waitFor, within } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import { CharacterValidationForm } from '../CharacterValidationForm';
-import { setupFormComponentTest } from './setup/test-setup';
-
-// Mock the CharacterService
-jest.mock('@/lib/services/CharacterService', () => ({
-  CharacterService: {
-    createCharacter: jest.fn(),
-  },
-}));
-
-import { CharacterService } from '@/lib/services/CharacterService';
-
-describe('CharacterValidationForm', () => {
-  setupFormComponentTest();
-  const mockCharacterService = CharacterService as jest.Mocked<typeof CharacterService>;
-
-  const defaultProps = {
-    ownerId: 'user123',
-    onSuccess: jest.fn(),
-    onError: jest.fn(),
-    onCancel: jest.fn(),
-    isOpen: false,
-  };
-
-  const testProps = {
-    ...defaultProps,
-    ownerId: 'test-owner-id',
-    isOpen: true,
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    // Set up default successful response
-    mockCharacterService.createCharacter.mockResolvedValue({
-      success: true,
-      data: {
-        _id: 'test-char-id',
-        name: 'Test Character',
-        type: 'pc',
-        race: 'human',
-        size: 'medium',
-      },
-    });
-  });
-
-  describe('Form Rendering', () => {
-    it('renders the form modal when open', () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      expect(screen.getByRole('heading', { name: 'Create Character' })).toBeInTheDocument();
-      expect(screen.getByText('Build your character with real-time validation')).toBeInTheDocument();
-    });
-
-    it('does not render when closed', () => {
-      render(<CharacterValidationForm {...testProps} isOpen={false} />);
-
-      expect(screen.queryByRole('heading', { name: 'Create Character' })).not.toBeInTheDocument();
-    });
-
-    it('renders all form sections', () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      expect(screen.getByTestId('basic-info-validation-section')).toBeInTheDocument();
-      expect(screen.getByTestId('ability-scores-validation-section')).toBeInTheDocument();
-      expect(screen.getByTestId('classes-validation-section')).toBeInTheDocument();
-      expect(screen.getByTestId('combat-stats-validation-section')).toBeInTheDocument();
-    });
-  });
-
-  describe('Real-time Validation', () => {
-    it('displays validation errors for empty required fields', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-
-      // First type something to trigger validation, then clear it
-      await userEvent.type(nameInput, 'Test');
-      await userEvent.clear(nameInput);
-      await userEvent.tab(); // Trigger blur to show validation
-
-      await waitFor(() => {
-        expect(screen.getByText(/Name is required/)).toBeInTheDocument();
-      });
-    });
-
-    it('shows form status as invalid when there are errors', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.clear(nameInput);
-      await userEvent.tab();
-
-      await waitFor(() => {
-        expect(screen.getByText('Form Status: Invalid')).toBeInTheDocument();
-      });
-    });
-
-    it('shows form status as valid when all fields are correct', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Valid Character Name');
-
-      await waitFor(() => {
-        expect(screen.getByText('Form Status: Valid')).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Basic Info Section', () => {
-    it('updates character name field', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Aragorn');
-
-      expect(nameInput).toHaveValue('Aragorn');
-      expect(screen.getByText('7/100')).toBeInTheDocument(); // Character count
-    });
-
-    it('handles custom race input', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Find the race select trigger by role
-      const raceSelect = screen.getByRole('combobox', { name: /Race/ });
-      await userEvent.click(raceSelect);
-
-      // Wait for dropdown to open and click Custom option
-      await waitFor(() => {
-        expect(screen.getByText('Custom')).toBeInTheDocument();
-      });
-      await userEvent.click(screen.getByText('Custom'));
-
-      // Should show custom race input
-      await waitFor(() => {
-        expect(screen.getByLabelText(/Custom Race Name/)).toBeInTheDocument();
-      });
-
-      const customRaceInput = screen.getByLabelText(/Custom Race Name/);
-      await userEvent.type(customRaceInput, 'Dragonkin');
-
-      expect(customRaceInput).toHaveValue('Dragonkin');
-    });
-
-    it('updates size field', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Find the size select trigger by role
-      const sizeSelect = screen.getByRole('combobox', { name: /Size/ });
-      await userEvent.click(sizeSelect);
-
-      // Wait for dropdown to open and click Large option
-      await waitFor(() => {
-        expect(screen.getByText('Large')).toBeInTheDocument();
-      });
-      await userEvent.click(screen.getByText('Large'));
-
-      // Verify the selection was made by checking the trigger shows Large
-      await waitFor(() => {
-        expect(screen.getByText('Large')).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Ability Scores Section', () => {
-    it('updates ability scores and shows modifiers', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const strengthInput = screen.getByLabelText(/Strength \(STR\)/);
-
-      // The input field should have the default value
-      expect(strengthInput).toHaveValue(10);
-
-      // Should show modifiers for all ability scores (all start at 10, so all are +0)
-      const modifiers = screen.getAllByText('+0');
-      expect(modifiers).toHaveLength(6); // 6 ability scores, all with +0 modifier
-    });
-
-    it('validates ability score ranges', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const strengthInput = screen.getByLabelText(/Strength \(STR\)/);
-
-      // Verify the input accepts valid values
-      expect(strengthInput).toHaveValue(10);
-
-      // Check that the form status section exists
-      expect(screen.getByText(/Form Status:/)).toBeInTheDocument();
-    });
-  });
-
-  describe('Classes Section', () => {
-    it('adds and removes character classes', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Should start with one class
-      expect(screen.getByText('Class 1 (Primary)')).toBeInTheDocument();
-
-      // Add a second class
-      await userEvent.click(screen.getByText('Add Class'));
-
-      await waitFor(() => {
-        expect(screen.getByText('Class 2')).toBeInTheDocument();
-      });
-
-      // Remove the second class - look for button with X icon and sr-only text
-      // Only non-primary classes should have remove buttons
-      const removeButtons = screen.getAllByRole('button').filter(button =>
-        button.querySelector('.sr-only')?.textContent === 'Remove class'
-      );
-      // Now that we have 2 classes, both should have remove buttons (component logic changed)
-      expect(removeButtons.length).toBeGreaterThan(0);
-      await userEvent.click(removeButtons[removeButtons.length - 1]); // Click the last remove button
-
-      await waitFor(() => {
-        expect(screen.queryByText('Class 2')).not.toBeInTheDocument();
-      });
-    });
-
-    it('auto-updates hit die when class changes', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Find the character class select by label
-      const classSelect = screen.getByRole('combobox', { name: /Character Class/ });
-      await userEvent.click(classSelect);
-
-      // Wait for dropdown to open and click Barbarian option
-      await waitFor(() => {
-        expect(screen.getByText('Barbarian')).toBeInTheDocument();
-      });
-      await userEvent.click(screen.getByText('Barbarian'));
-
-      // Hit die should auto-update to 12 for barbarian
-      await waitFor(() => {
-        const hitDieInput = screen.getByLabelText(/Hit Die/);
-        expect(hitDieInput).toHaveValue(12);
-      });
-    });
-
-    it('prevents adding more than 3 classes', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Add maximum classes (starts with 1, add 2 more to reach 3)
-      const addButton = screen.getByRole('button', { name: /Add Class/ });
-      await userEvent.click(addButton);
-      await userEvent.click(addButton);
-
-      // Button should be disabled after reaching 3 classes
-      await waitFor(() => {
-        const button = screen.getByRole('button', { name: /Add Class/ });
-        expect(button).toBeDisabled();
-      });
-
-      // Should show 3 classes
-      expect(screen.getByText('Class 1 (Primary)')).toBeInTheDocument();
-      expect(screen.getByText('Class 2')).toBeInTheDocument();
-      expect(screen.getByText('Class 3')).toBeInTheDocument();
-    });
-  });
-
-  describe('Combat Stats Section', () => {
-    it('updates hit points values', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Wait for the form to render
-      await waitFor(() => {
-        expect(screen.getByTestId('combat-stats-validation-section')).toBeInTheDocument();
-      });
-
-      const maxHpInput = screen.getByLabelText(/Maximum HP/);
-
-      // Clear field by selecting all and typing new value
-      await userEvent.tripleClick(maxHpInput);
-      await userEvent.keyboard('15');
-
-      expect(maxHpInput).toHaveValue(15);
-
-      const currentHpInput = screen.getByLabelText(/Current HP/);
-      await userEvent.tripleClick(currentHpInput);
-      await userEvent.keyboard('12');
-
-      expect(currentHpInput).toHaveValue(12);
-    });
-
-    it('shows combat summary', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Wait for the form to render and check for combat summary
-      await waitFor(() => {
-        expect(screen.getByText('Combat Summary')).toBeInTheDocument();
-      });
-
-      // Get the parent container that includes both title and stats
-      const combatSummaryContainer = screen.getByText('Combat Summary').closest('.p-4');
-      expect(combatSummaryContainer).toBeInTheDocument();
-
-      // Look for the specific combat summary stats within the container
-      within(combatSummaryContainer!).getByText(/HP:/);
-      within(combatSummaryContainer!).getByText(/AC:/);
-      within(combatSummaryContainer!).getByText(/Speed:/);
-      within(combatSummaryContainer!).getByText(/Prof:/);
-    });
-  });
-
-  describe('Form Submission', () => {
-    it('submits valid form data', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Fill in required fields
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Test Character');
-
-      // Submit the form
-      await userEvent.click(screen.getByRole('button', { name: 'Create Character' }));
-
-      await waitFor(() => {
-        expect(mockCharacterService.createCharacter).toHaveBeenCalledWith(
-          'test-owner-id',
-          expect.objectContaining({
-            name: 'Test Character',
-            type: 'pc',
-            race: 'human',
-            size: 'medium',
-          })
-        );
-      });
-
-      expect(testProps.onSuccess).toHaveBeenCalled();
-    });
-
-    it('displays error message on submission failure', async () => {
-      mockCharacterService.createCharacter.mockResolvedValue({
-        success: false,
-        error: {
-          code: 'VALIDATION_ERROR',
-          message: 'Invalid character data',
-          details: 'Name is required',
-        },
-      });
-
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Test Character');
-
-      await userEvent.click(screen.getByRole('button', { name: 'Create Character' }));
-
-      await waitFor(() => {
-        expect(screen.getByText(/Failed to create character: Invalid character data/)).toBeInTheDocument();
-        expect(screen.getByText(/Details: Name is required/)).toBeInTheDocument();
-      });
-    });
-
-    it('prevents submission when form is invalid', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Clear required field to make form invalid
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.clear(nameInput);
-
-      // Submit button should be disabled or not functional
-      await userEvent.click(screen.getByRole('button', { name: 'Create Character' }));
-
-      // Should not call the service
-      expect(mockCharacterService.createCharacter).not.toHaveBeenCalled();
-    });
-
-    it('shows loading state during submission', async () => {
-      // Mock a delayed response
-      mockCharacterService.createCharacter.mockImplementation(
-        () => new Promise(resolve => setTimeout(() => resolve({
-          success: true,
-          data: {
-            _id: 'test-character-id',
-            name: 'Test Character',
-            type: 'pc',
-            race: 'human',
-            size: 'medium',
-            classes: [{ class: 'fighter', level: 1, hitDie: 10 }],
-            abilityScores: {
-              strength: 15,
-              dexterity: 14,
-              constitution: 13,
-              intelligence: 12,
-              wisdom: 10,
-              charisma: 8,
-            },
-            hitPoints: { maximum: 10, current: 10, temporary: 0 },
-            armorClass: 16,
-            speed: 30,
-            proficiencyBonus: 2,
-            savingThrows: {
-              strength: false,
-              dexterity: false,
-              constitution: false,
-              intelligence: false,
-              wisdom: false,
-              charisma: false,
-            },
-            skills: {},
-            equipment: [],
-            spells: [],
-            ownerId: 'test-owner-id',
-            createdAt: new Date().toISOString(),
-            updatedAt: new Date().toISOString(),
-          },
-        }), 100))
-      );
-
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Test Character');
-
-      const submitButton = screen.getByRole('button', { name: 'Create Character' });
-      await userEvent.click(submitButton);
-
-      // Should show loading state immediately
-      await waitFor(() => {
-        expect(screen.getByRole('button', { name: /Creating/ })).toBeInTheDocument();
-      }, { timeout: 3000 });
-    });
-  });
-
-  describe('Character Preview', () => {
-    it('updates preview when form values change', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      const nameInput = screen.getByLabelText(/Character Name/);
-      await userEvent.type(nameInput, 'Preview Character');
-
-      // Preview should update with the new name
-      await waitFor(() => {
-        expect(screen.getByText('Preview Character')).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Modal Behavior', () => {
-    it('calls onCancel when cancel button is clicked', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      await userEvent.click(screen.getByText('Cancel'));
-
-      expect(defaultProps.onCancel).toHaveBeenCalled();
-    });
-
-    it('calls onCancel when modal is closed', async () => {
-      render(<CharacterValidationForm {...testProps} />);
-
-      // Find the modal's close button (usually an X button) or press Escape
-      // The Modal should call onOpenChange with false when closed
-      const user = userEvent.setup();
-      await user.keyboard('{Escape}');
-
-      await waitFor(() => {
-        expect(defaultProps.onCancel).toHaveBeenCalled();
-      });
-    });
-  });
-
-  describe('Initial Values', () => {
-    it('pre-populates form with initial values', async () => {
-      const initialValues = {
-        name: 'Pre-filled Character',
-        type: 'npc' as const,
-        race: 'elf' as const,
-        size: 'small' as const,
-      };
-
-      render(
-        <CharacterValidationForm
-          {...testProps}
-          initialValues={initialValues}
-        />
-      );
-
-      // Check the name input field
-      expect(screen.getByDisplayValue('Pre-filled Character')).toBeInTheDocument();
-
-      // For Select components, check that the form rendered successfully
-      expect(screen.getByTestId('basic-info-validation-section')).toBeInTheDocument();
-
-      // The select values might not be immediately visible as text
-      // Just verify the form rendered with the initial values successfully
-      const form = screen.getByRole('dialog');
-      expect(form).toBeInTheDocument();
-    });
-  });
-});
+// Main test entry point for CharacterValidationForm
+// Tests have been split into focused files for better maintainability:
+
+import './CharacterValidationForm.rendering.test';
+import './CharacterValidationForm.validation.test';
+import './CharacterValidationForm.sections.test';
+import './CharacterValidationForm.classes.test';
+import './CharacterValidationForm.submission.test';
+import './CharacterValidationForm.modal.test';
+
+// Note: This file serves as a centralized entry point for all CharacterValidationForm tests.
+// Each test file focuses on a specific aspect of the component:
+//
+// - rendering.test: Basic rendering and initial values
+// - validation.test: Real-time validation and form status
+// - sections.test: Basic info, ability scores, combat stats, and character preview
+// - classes.test: Class management (add, remove, limits)
+// - submission.test: Form submission, error handling, loading states
+// - modal.test: Modal behavior and interactions
+//
+// This organization keeps each file under the 450-line limit while maintaining
+// comprehensive test coverage for the CharacterValidationForm component.
